+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.2] - 2024-03-26
+
+### Added
+
+- [MatterbridgeDevice]: Added Cluster DoorLock and command handler.
+
 ## [1.2.1] - 2024-03-25
 
 ### Added